Commit #23
Now instead of having "game_won_fragment" and "game_lost_fragment" everything is shown on "result_fragment". The logic for thar was improved but and now the app is more eficient.

The user can also share the results in any case.

On a later upgrade I will make sure the result is shareable only in case of wins